+++ src/index.js
@@ -20,6 +20,7 @@ export { isTriangular } from "./combinatorics/isTriangular.js";
 export { lucas } from "./combinatorics/lucas.js";
 export { nthFibonacci } from "./combinatorics/nthFibonacci.js";
 export { nthLucas } from "./combinatorics/nthLucas.js";
+export { pentagonal } from "./combinatorics/pentagonal.js";
 export { permutation } from "./combinatorics/permutation.js";
 export { square } from "./combinatorics/square.js";
 export { triangular } from "./combinatorics/triangular.js";